Refactor auth flow and unify response DTOs
Replaces RegisterResponseDto with AuthResponseDto for both registration and login, and implements login endpoint in AuthController. Updates User entity and related methods to use Long for userId. Refactors JwtUtil and repository methods to support Long userId, and adds login logic to AuthService.

+++ finanzas-back/src/main/kotlin/com/finanzas/dash/finanzas/config/security/JwtUtil.kt
@@ -13,7 +13,7 @@ class JwtUtil(private val jwtConfigValues: JwtConfigValues) {
     private val key: SecretKey = Keys.hmacShaKeyFor(jwtConfigValues.secret.toByteArray())
     private val expirationMs: Long = jwtConfigValues.expiration * 15 // 15 minutos
 
-    fun generateToken(id: Int): String {
+    fun generateToken(id: Long): String {
         return Jwts.builder()
             .subject(id.toString())
             .issuedAt(Date())
